+++ src/constellation/__init__.py
@@ -2,4 +2,4 @@
 
 __all__ = ["__version__"]
 
-__version__ = "6.1.0"
+__version__ = "6.1.1"